// apps/web/lib/types.ts
export type CustomXYInputValue = {
  top?: number;
  right?: number;
  bottom?: number;
  left?: number;
}

export type CustomXYInputValues = {
  mobile?: CustomXYInputValue;
  tablet?: CustomXYInputValue;
  desktop?: CustomXYInputValue;
};

export type CustomStyleValues = {
  mobile?: string;
  tablet?: string;
  desktop?: string;
};